:recycle: Refactor headerV2 code
