+++ i18n/locales/en/common.json
@@ -482,7 +482,6 @@
   "Try using menu for navigation.": "Try using menu for navigation.",
   "Type": "Type",
   "Type at least 3 characters": "Type at least 3 characters",
-  "Unable to connect to the node": "Unable to connect to the node",
   "Unable to connect to the node, no response from the server.": "Unable to connect to the node, no response from the server.",
   "Unable to connect to the node, please check the address and try again": "Unable to connect to the node, please check the address and try again",
   "Unable to detect the communication layer with your Hardware Wallet": "Unable to detect the communication layer with your Hardware Wallet",
