+++ src/components/headerV2/index.js
@@ -1,23 +1,16 @@
 /* istanbul ignore file */
 import { connect } from 'react-redux';
-import { setDefaults, translate } from 'react-i18next';
-import LoginV2 from './loginV2';
+import HeaderV2 from './headerV2';
 import { liskAPIClientSet } from '../../actions/peers';
 import { errorToastDisplayed } from '../../actions/toaster';
 import { settingsUpdated } from '../../actions/settings';
 
-setDefaults({
-  wait: true,
-  withRef: false,
-  bindI18n: 'languageChanged loaded',
-  bindStore: 'added removed',
-  nsMode: 'default',
-  translateFuncName: 't',
-});
-
 const mapStateToProps = state => ({
   account: state.account,
+  peers: state.peers,
   settings: state.settings,
+  selectedNetwork: (state.peers && state.peers.options.code) || 0,
+  address: state.peers && state.peers.options.address,
   liskAPIClient: state.peers && state.peers.liskAPIClient,
 });
 
@@ -30,4 +23,4 @@ const mapDispatchToProps = {
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(translate()(LoginV2));
+)(HeaderV2);
